Add GlobalOptions_InProc/OutOfProc and move option helpers to the new type
