+++ src/VisualStudio/IntegrationTest/TestUtilities/OutOfProcess/VisualStudioWorkspace_OutOfProc.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
+using System.Collections.Immutable;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
 using Microsoft.CodeAnalysis.SolutionCrawler;
@@ -20,12 +21,6 @@ namespace Microsoft.VisualStudio.IntegrationTest.Utilities.OutOfProcess
             _inProc = CreateInProcComponent<VisualStudioWorkspace_InProc>(visualStudioInstance);
         }
 
-        public bool IsPrettyListingOn(string languageName)
-            => _inProc.IsPrettyListingOn(languageName);
-
-        public void SetPrettyListing(string languageName, bool value)
-            => _inProc.SetPrettyListing(languageName, value);
-
         public void WaitForAsyncOperations(TimeSpan timeout, string featuresToWaitFor, bool waitForWorkspaceFirst = true)
             => _inProc.WaitForAsyncOperations(timeout, featuresToWaitFor, waitForWorkspaceFirst);
 
@@ -38,33 +33,7 @@ namespace Microsoft.VisualStudio.IntegrationTest.Utilities.OutOfProcess
         public void CleanUpWorkspace()
             => _inProc.CleanUpWorkspace();
 
-        public void ResetOptions()
-            => _inProc.ResetOptions();
-
         public void CleanUpWaitingService()
             => _inProc.CleanUpWaitingService();
-
-        public void SetImportCompletionOption(bool value)
-        {
-            SetGlobalOption(WellKnownGlobalOption.CompletionOptions_ShowItemsFromUnimportedNamespaces, LanguageNames.CSharp, value);
-            SetGlobalOption(WellKnownGlobalOption.CompletionOptions_ShowItemsFromUnimportedNamespaces, LanguageNames.VisualBasic, value);
-        }
-
-        public void SetEnableDecompilationOption(bool value)
-        {
-            SetGlobalOption(WellKnownGlobalOption.MetadataAsSourceOptions_NavigateToDecompiledSources, language: null, value);
-        }
-
-        public void SetArgumentCompletionSnippetsOption(bool value)
-        {
-            SetGlobalOption(WellKnownGlobalOption.CompletionViewOptions_EnableArgumentCompletionSnippets, LanguageNames.CSharp, value);
-            SetGlobalOption(WellKnownGlobalOption.CompletionViewOptions_EnableArgumentCompletionSnippets, LanguageNames.VisualBasic, value);
-        }
-
-        public void SetTriggerCompletionInArgumentLists(bool value)
-            => SetGlobalOption(WellKnownGlobalOption.CompletionOptions_TriggerInArgumentLists, LanguageNames.CSharp, value);
-
-        public void SetGlobalOption(WellKnownGlobalOption option, string? language, object? value)
-            => _inProc.SetGlobalOption(option, language, value);
     }
 }
